added test with d128 as key and moved the util function
