+++ integration_tests/src/main/python/map_test.py
@@ -342,10 +342,21 @@ def test_element_at_map_numeric_keys(data_gen):
         conf={'spark.sql.ansi.enabled': False})
 
 @pytest.mark.parametrize('data_gen',
+                         [MapGen(DecimalGen(precision=35, scale=2, nullable=False), value(), max_length=6)
+                          for value in get_map_value_gens(precision=37, scale=0)],
+                         ids=idfn)
+def test_get_map_value_element_at_map_dec_col_keys(data_gen):
+    keys = DecimalGen(precision=35, scale=2)
+    assert_gpu_and_cpu_are_equal_collect(
+        lambda spark: two_col_df(spark, data_gen, keys).selectExpr(
+            'element_at(a, b)', 'a[b]'),
+        conf={'spark.sql.ansi.enabled': False})
+
+@pytest.mark.parametrize('data_gen',
                          [MapGen(StringGen(pattern='key_[0-9]', nullable=False), value(), max_length=6)
                           for value in get_map_value_gens(precision=37, scale=0)],
                          ids=idfn)
-def test_element_at_map_string_col_keys(data_gen):
+def test_get_map_value_element_at_map_string_col_keys(data_gen):
     keys = StringGen(pattern='key_[0-9]')
     assert_gpu_and_cpu_are_equal_collect(
         lambda spark: two_col_df(spark, data_gen, keys).selectExpr(
